Split services tests
